+++ auth/iam/metadata.go
@@ -2,11 +2,6 @@ package iam
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
-	"net"
-	"net/http"
 	"sync"
 	"time"
 
@@ -16,99 +11,105 @@ import (
 var _ ydb.Credentials = &instanceServiceAccountCredentials{}
 
 type instanceServiceAccountCredentials struct {
-	token     string
-	expiresAt time.Time
+	mu  *sync.RWMutex
+	ctx context.Context
 
-	mu *sync.RWMutex
+	token string
+	err   error
 
-	httpClient http.Client
+	expiry time.Time
+	timer  *time.Timer
 }
 
 // Returns cached token if it is valid. Otherwise, will try to renew.
-func (m *instanceServiceAccountCredentials) Token(ctx context.Context) (string, error) {
-	m.mu.RLock()
-
-	const renewOverhead = 15 * time.Second
-	if m.token != "" {
-		if time.Until(m.expiresAt) > renewOverhead {
-			defer m.mu.RUnlock()
-			return m.token, nil
+func (m *instanceServiceAccountCredentials) Token(ctx context.Context) (token string, err error) {
+	for {
+		select {
+		case <-ctx.Done():
+			return "", &CreateTokenError{
+				Cause:  ctx.Err(),
+				Reason: ctx.Err().Error(),
+			}
+		default:
+			m.mu.RLock()
+			token, err = m.token, m.err
+			m.mu.RUnlock()
+			if token != "" || err != nil {
+				return token, err
+			}
+			// not yet initialized, wait
 		}
 	}
+}
 
-	//renew
-	m.mu.RUnlock()
+func (m *instanceServiceAccountCredentials) refreshLoop() {
+	defer m.timer.Stop()
+	for {
+		select {
+		case <-m.ctx.Done():
+			// Set up error
+			m.mu.Lock()
+			m.token, m.err = "", &CreateTokenError{
+				Cause:  m.ctx.Err(),
+				Reason: m.ctx.Err().Error(),
+			}
+			m.mu.Unlock()
+			return
+		case <-m.timer.C:
+			m.refreshOnce()
+		}
+	}
+}
+
+// Perform single refresh iteration.
+// If token was obtained:
+// 1. Clear current err;
+// 2. Set up new token and expiration;
+// Otherwise, if current token has expired, clear it and set up err.
+func (m *instanceServiceAccountCredentials) refreshOnce() {
+	now := time.Now()
+	tok, err := metaCall()
+
+	// Call has been performed, now updating fields
 	m.mu.Lock()
 	defer m.mu.Unlock()
 
-	const metadataURL = "http://169.254.169.254/computeMetadata/v1/instance/service-accounts/default/token" // from YC go-sdk
-	req, err := http.NewRequest("GET", metadataURL, nil)
+	defer func() {
+		const minInterval = 5 * time.Second
+		// Reset timer: trigger after 10% of expiry.
+		// NB: we are guaranteed to have drained timer here.
+		interval := time.Until(m.expiry) / 10
+		if interval < minInterval {
+			interval = minInterval
+		}
+		m.timer.Reset(interval)
+	}()
+
 	if err != nil {
-		return "", &CreateTokenError{
-			Reason: fmt.Errorf("failed to create HTTP request"),
+		// Check if current value is still good.
+		if m.expiry.After(now) {
+			// Will leave old token in place
+			return
 		}
+		// Clear token and set up err
+		m.token = ""
+		m.err = err
+		return
 	}
-
-	req.Header.Set("Metadata-Flavor", "Google") // from YC go-sdk
-
-	reqTime := time.Now()
-	resp, err := m.httpClient.Do(req.WithContext(ctx))
-	if err != nil {
-		return "", &CreateTokenError{
-			Reason: fmt.Errorf("failed medatata request (outside of compute instance?): %s",
-				err,
-			),
-		}
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode == http.StatusNotFound {
-		return "", &CreateTokenError{
-			Reason: fmt.Errorf("%s: possibly missing service_account_id in instance spec",
-				resp.Status,
-			),
-		}
-	}
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return "", &CreateTokenError{
-			Reason: fmt.Errorf("response body read failed: %s", err.Error()),
-		}
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		return "", fmt.Errorf("%s", resp.Status)
-	}
-
-	var tokenResponse struct {
-		AccessToken string `json:"access_token"`
-		ExpiresIn   int64  `json:"expires_in"`
-		TokenType   string `json:"token_type"`
-	}
-
-	err = json.Unmarshal(body, &tokenResponse)
-	if err != nil {
-		return "", &CreateTokenError{
-			Reason: fmt.Errorf("failed response body unmarshal: %s", err.Error()),
-		}
-	}
-	m.token = tokenResponse.AccessToken
-	m.expiresAt = reqTime.Add(time.Duration(tokenResponse.ExpiresIn) * time.Second)
-
-	return m.token, nil
+	// Renew values.
+	m.token, m.expiry, m.err = tok.Token, now.Add(tok.ExpiresIn), nil
 }
 
 // Credentials provider that uses instance metadata to obtain token for service account attached to instance.
-func InstanceServiceAccount() (ydb.Credentials, error) {
-	return &instanceServiceAccountCredentials{
-		mu: &sync.RWMutex{},
-		httpClient: http.Client{
-			Transport: &http.Transport{
-				DialContext: (&net.Dialer{
-					Timeout:   time.Second, // One second should be enough for localhost connection.
-					KeepAlive: -1,          // No keep alive. Near token per hour requested.
-				}).DialContext,
-			},
-		},
-	}, nil
+// Cancelling context will lead to credentials refresh halt.
+// It should be used during application stop or credentials recreation.
+func InstanceServiceAccount(ctx context.Context) ydb.Credentials {
+	credentials := &instanceServiceAccountCredentials{
+		mu:    &sync.RWMutex{},
+		ctx:   ctx,
+		timer: time.NewTimer(0), // Allocate expired
+	}
+	// Start refresh loop.
+	go credentials.refreshLoop()
+	return credentials
 }